pyproject.toml
[tool.poetry]
name = "ecigs-vape"
version = "0.1.0"
description = ""
authors = ["Oleh Salitrynskyi"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2.3"
python-decouple = "^3.8"
django-dirtyfields = "^1.9.2"
Pillow = "^10.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
